drop the old working variants merge the latest variant without metalb and new route53 into the last working i.e. single variant left, as s7, under trial
