+++ .github/workflows/rke-nginx-exts3.yml
@@ -17,7 +17,7 @@ on:
     branches: [ main ]
     types: [ ready_for_review ]
     paths:
-      - 'acceptance/install/scenario8_test.go'
+      - 'acceptance/install/scenario7_test.go'
       - '.github/workflows/rke-nginx-exts3.yml'
   schedule:
     - cron:  '0 2 * * *'
@@ -116,7 +116,7 @@ jobs:
         with:
           username: ${{ secrets.DOCKERHUB_USERNAME }}
           password: ${{ secrets.DOCKERHUB_TOKEN }}
-          
+
       - name: Install Tools
         run: make tools-install
 
@@ -145,7 +145,7 @@ jobs:
 
       - name: Installation Acceptance Tests
         env:
-          REGEX: Scenario8
+          REGEX: Scenario7
           PUBLIC_CLOUD: 1
           KUBECONFIG: /etc/rancher/rke2/rke2.yaml
           REGISTRY_USERNAME: ${{ secrets.CFCIBOT_DOCKERHUB_USERNAME }}
@@ -154,7 +154,7 @@ jobs:
           AWS_ZONE_ID: ${{ github.event.inputs.aws_zone_id || secrets.AWS_ZONE_ID }}
           AWS_ACCESS_KEY_ID: ${{ github.event.inputs.aws_id || secrets.AWS_ACCESS_KEY_ID }}
           AWS_SECRET_ACCESS_KEY: ${{ github.event.inputs.aws_key || secrets.AWS_SECRET_ACCESS_KEY }}
-	  AWS_BASE_DOMAIN: ${{ github.event.inputs.aws_domain || secrets.AWS_DOMAIN }}
+          AWS_BASE_DOMAIN: ${{ github.event.inputs.aws_domain || secrets.AWS_DOMAIN }}
           # EXTRAENV_NAME: SESSION_KEY
           # EXTRAENV_VALUE: 12345
         run: |
